chore: file restructure - user settings
