+++ frontend/src/components/common/UserAvatar.tsx
@@ -5,7 +5,7 @@ import { clsx } from 'clsx'
 import { generateAvatarColor } from '../feature/select-member/GenerateAvatarColor'
 import { RiRobot2Fill } from 'react-icons/ri'
 import { useMemo } from 'react'
-import { AvailabilityStatus } from '../feature/userSettings/SetUserAvailabilityMenu'
+import { AvailabilityStatus } from '../feature/userSettings/AvailabilityStatus/SetUserAvailabilityMenu'
 
 interface UserAvatarProps extends Partial<AvatarProps> {
     alt?: string,
