+++ frontend/src/components/feature/userSettings/UploadImage/ImageUploader.tsx
@@ -6,7 +6,7 @@ import { DIALOG_CONTENT_CLASS } from "@/utils/layout/dialog"
 import { DeleteImageModal } from "./DeleteImageModal"
 import { useUserData } from "@/hooks/useUserData"
 import { toast } from "sonner"
-import { CustomFile } from "../file-upload/FileDrop"
+import { CustomFile } from "../../file-upload/FileDrop"
 import { UploadImageModal } from "./UploadImageModal"
 import { BiTrash } from "react-icons/bi"
 import { LuUpload } from "react-icons/lu"
