+++ frontend/src/components/feature/userSettings/UserSettings.tsx
@@ -2,6 +2,7 @@ import { PageHeader } from '@/components/layout/Heading/PageHeader'
 import { Box, Flex, Heading, Text } from '@radix-ui/themes'
 import { BiChevronLeft } from 'react-icons/bi'
 import { Link } from 'react-router-dom'
+import { ImageUploader } from './UploadImage/ImageUploader'
 
 const UserSettings = () => {
     return (
@@ -17,6 +18,7 @@ const UserSettings = () => {
             <Box className="min-h-screen pt-16 pb-8">
                 <Flex direction='column' gap='3' justify='start' px='4'>
                     <Text size="2">Make changes to your account.</Text>
+                    <ImageUploader />
                 </Flex>
             </Box>
         </>
